# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B2B Expense & Team Budget Tracker",
    "version": "1.0.0",
    "summary": "Expense claims, department budgets and recurring vendor spend for multi-tenant B2B customers.",
    "description": "Tenancy is enforced by PostgreSQL row-level security, not by application filters. The tenant travels only in the signed access token: there is no tenant header, path segment or query parameter, because any of those would be a value the client chooses.\\n\\nMoney is always an integer count of minor units plus a currency. No amount is ever a JSON number with a decimal point: JSON numbers are IEEE 754 doubles in every browser, and a system that sums thousands of expense lines that way produces totals that disagree with the receipts.\\n\\nErrors share one envelope. The status code is chosen for what the client should do next: 401 log in, 402 change plan, 403 ask someone else, 404 stop asking, 409 reload and retry, 413 narrow the range, 422 fix a field, 503 retry later.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/reports/expenses/export": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Stream an expense report as CSV, XLSX or PDF",
        "operationId": "getReportsExpensesExport",
        "description": "Streams the report. Nothing is buffered: rows are encoded and written as they are read.\\n\\nAuthorised by either a bearer token or the `token` query parameter from `/reports/expenses/export/token`. The signed token exists because a browser navigating to a download cannot set a header, and it is bound to the exact query it was issued for.",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "csv",
                "xlsx",
                "pdf"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/FilterStatus"
          },
          {
            "$ref": "#/components/parameters/FilterDepartment"
          },
          {
            "$ref": "#/components/parameters/FilterFrom"
          },
          {
            "$ref": "#/components/parameters/FilterTo"
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "A signed download token, when no bearer token is presented."
          }
        ],
        "responses": {
          "200": {
            "description": "The report, streamed",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "description": "The range exceeds the synchronous row limit for this plan."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A short-lived access token carrying the user and the tenant. The role is deliberately not in the token: a token issued before a demotion would otherwise keep its old authority until it expired."
      },
      "relaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Billing-Signature",
        "description": "HMAC-SHA256 over the timestamp and the raw body, verified before the event id is claimed."
      }
    },
    "parameters": {
...
      "FilterStatus": {
        "name": "status",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/ExpenseStatus"
        }
      },
...
      "FilterDepartment": {
        "name": "department_id",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "FilterFrom": {
        "name": "from",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date"
        },
        "description": "Inclusive lower bound on spent_at - the date on the receipt, not the date the claim was filed."
      },
      "FilterTo": {
        "name": "to",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date"
        }
      },
...
    },
    "schemas": {
...
      "ExpenseStatus": {
        "type": "string",
        "enum": [
          "draft",
          "pending_approval",
          "approved",
          "rejected",
          "paid"
        ]
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string",
            "description": "Withheld for 5xx: an internal failure's text can name a table, a constraint or a host."
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "trace_id": {
            "type": "string",
            "description": "Matches X-Request-Id, and appears on every log line for the request."
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "detail"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid credential. Every rejected token gets the same answer: distinguishing expired from malformed from wrongly-signed tells a forger which part to fix next.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated, but not permitted. Ask someone else - this will not resolve by retrying.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such resource, or none you may see. The two are deliberately indistinguishable: a 403 on a resource outside your scope would confirm the id exists.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Invalid": {
        "description": "Well formed, but the values are wrong. The fields array names each one.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}